[gamedig] Fix attw errors for 5.0 stable
